requirements.txt
apache-airflow==2.10.4
apache-airflow-providers-google==10.22.0
pyspark==3.5.3
requests==2.32.3
google-cloud-storage==2.18.2
google-cloud-bigquery==3.26.0
pandas==2.2.3
pyarrow==17.0.0
fastparquet==2024.6.3
numpy==1.26.4